+++ designs/io_placer/src/user_proj_example2.v
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // SPDX-License-Identifier: Apache-2.0
-
+/// sta-blackbox
 `default_nettype none
 /*
  *-------------------------------------------------------------