+++ openlane/scripts/openroad/rsz_timing_postcts.tcl
@@ -18,7 +18,7 @@ set_propagated_clock [all_clocks]
 
 # set don't touch nets
 source $::env(SCRIPTS_DIR)/openroad/common/resizer.tcl
-set_dont_touch_rx "$::env(RSZ_DONT_TOUCH_RX)"
+set_dont_touch_objects
 
 # set don't use cells
 if { [info exists ::env(RSZ_DONT_USE_CELLS)] } {
@@ -55,7 +55,7 @@ if { [catch {check_placement -verbose} errmsg] } {
     exit 1
 }
 
-unset_dont_touch_rx "$::env(RSZ_DONT_TOUCH_RX)"
+unset_dont_touch_objects
 
 source $::env(SCRIPTS_DIR)/openroad/common/set_rc.tcl
 estimate_parasitics -placement