+++ Changelog.md
@@ -1,3 +1,12 @@
+# 2.0.0-a42
+* Added support for instances to `RSZ_DONT_TOUCH_RX`
+* Added support for `RSZ_DONT_TOUCH_LIST` to resizer steps
+* `inverter` design used to configure the above two
+* Added ignore for `//` key in JSON config files
+* Added two new variables for `Yosys.Synthesis`; `SYNTH_DIRECT_WIRE_BUFFERING` and `SYNTH_SPLITNETS`
+* Fixed bug with Yosys report parsing
+* Fixed issue in `usb_cdc_core` masked by aforementioned bug
+
 # 2.0.0-a41
 * Updated Magic to `9b131fa`
 * Updated Magic LEF writing script